requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.7.0
numpy>=1.24.0
scikit-image>=0.21.0
albumentations>=1.3.0
pycocotools>=2.0.6
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
Pillow>=9.5.0
requests>=2.31.0